Add convenience method to authenticate with app installation tokens; Convert existing markdown doc to APT due to formatting features; Replace occurrences of kohsuke user to github-api org where applicable; Fix repositoryIds method on GHAppCreateTokenBuilder;

+++ src/main/java/org/kohsuke/github/GHAppCreateTokenBuilder.java
@@ -35,7 +35,7 @@ public class GHAppCreateTokenBuilder {
      *
      */
     @Preview @Deprecated
-    public GHAppCreateTokenBuilder repositoryIds(List<Integer> repositoryIds) {
+    public GHAppCreateTokenBuilder repositoryIds(List<Long> repositoryIds) {
         this.builder.with("repository_ids",repositoryIds);
         return this;
     }
